Fix calling UserAction::update() for 'counters' only
This fixes:
> Undefined array key "data"

see 21dd1c3dafe4cfdc31748da4aae3027d63316d70

+++ wcfsetup/install/files/lib/data/user/UserAction.class.php
@@ -345,10 +345,9 @@ class UserAction extends AbstractDatabaseObjectAction implements IClipboardActio
             }
 
             if (
-                \array_key_exists(
-                    'password',
-                    $this->parameters['data']
-                ) && $this->parameters['data']['password'] !== ''
+                isset($this->parameters['data'])
+                && \array_key_exists('password', $this->parameters['data'])
+                && $this->parameters['data']['password'] !== ''
             ) {
                 foreach ($this->getObjects() as $object) {
                     SessionHandler::getInstance()->deleteUserSessionsExcept(
